Refactored main script code from Bash to Python.
This refactored main script code represents a migration from `Bash` to `Python`, enhancing the efficiency and readability of the codebase. By rewriting the original Bash script in Python, the code gains the benefits of a more structured and versatile programming language. This transition aims to improve maintainability, extensibility, and cross-platform compatibility. The Python script retains the core functionality of the original Bash script while taking advantage of Python's extensive libraries and cleaner syntax.

Additionally, added the essential files, specifically the 'avbtool python script,' which is required for the proper functioning of the script. Furthermore, both the 'custom-signed vbmeta' and the 'Private RSA key' for boot signing are now open-source.

Signed-off-by: Abhijeet